+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Summernote toolbar dropdowns not opening due to Bootstrap 5 event delegation conflict
+- Content deleted when emojis are used in text widgets [#0007619](https://bugs.oxid-esales.com/view.php?id=7619)
 
 ## [6.0.1] - 2025-11-10